fix a json viewer bug
